feat. Create Trending component & small fixes on hashtag css
