+++ src/pages/Hashtag/style.js
@@ -11,58 +11,28 @@ export const Container = styled.div`
 
 export const Wrapper = styled.div`
   display: flex;
+  justify-content: center;
+  width: 100%;
 `;
 
 export const PostsColumn = styled.div`
   width: 100%;
   max-width: 611px;
+
+  @media (max-width: 800px) {
   display: flex;
   flex-direction: column;
   align-items: center;
+  }
 `;
 
 export const H2 = styled.h2`
-  margin-left: 17px;
   margin-top: 91px;
+  margin-bottom: 25px;
   font-family: "Oswald";
   font-style: normal;
-  font-weight: 700;
+  font-weight: bold;
   font-size: 33px;
   line-height: 49px;
   color: #ffffff;
 `;
-
-export const TrendingContainer = styled.aside`
-  width: 301px;
-  height: 406px;
-  margin: 140px 0 0 25px;
-  background-color: #171717;
-  border-radius: 16px;
-`;
-
-export const TitleContainer = styled.div`
-  width: 100%;
-  height: 61px;
-  border-bottom: 1px solid #484848; ;
-  margin: 0 0 22px 0;
-`;
-
-export const TrendingTitle = styled.h3`
-  font-family: "Oswald";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 27px;
-  line-height: 61px;
-  color: #ffffff;
-  margin-left: 16px;
-`;
-
-export const TrendingTag = styled.p`
-  font-family: 'Lato';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 19px;
-  line-height: 30px;
-  color: #ffffff;
-  margin-left: 16px;
-`
